AI_devs4/se01e04/code/pyproject.toml
[project]
name = "spk-automation"
version = "1.0.0"
description = "SPK Package Automation System"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=0.3.0",
    "openai>=1.0.0",
    "langchain>=0.1.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "fastapi>=0.104.0",
    "uvicorn>=0.24.0",
    "httpx>=0.25.0",
    "aiohttp>=3.9.0",
    "pyyaml>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["mcp_server", "agent", "core"]

[dependency-groups]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
